+++ src/main/java/com/nhlstenden/amazonsimulatie/models/Rack.java
@@ -75,32 +75,32 @@ public class Rack implements Object3D {
   }
 
   @Override
-  public double getX() {
+  public int getX() {
     return x;
   }
 
   @Override
-  public double getY() {
+  public int getY() {
     return y;
   }
 
   @Override
-  public double getZ() {
+  public int getZ() {
     return z;
   }
 
   @Override
-  public double getRotationX() {
+  public int getRotationX() {
     return 0;
   }
 
   @Override
-  public double getRotationY() {
+  public int getRotationY() {
     return 0;
   }
 
   @Override
-  public double getRotationZ() {
+  public int getRotationZ() {
     return 0;
   }
